# Enforce HTTP/1.1 pipeline response order