--- a/http/racing_server_test.ts
+++ b/http/racing_server_test.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-const { run } = Deno;
+const { dial, run } = Deno;
 
 import { test } from "../testing/mod.ts";
 import { assert, assertEquals } from "../testing/asserts.ts";
@@ -23,13 +23,12 @@ function killServer(): void {
   server.stdout.close();
 }
 
-let nc;
-let ncIn = `GET / HTTP/1.1
+let input = `GET / HTTP/1.1
 
 GET / HTTP/1.1
 
 `;
-let ncOut = `HTTP/1.1 200 OK
+let output = `HTTP/1.1 200 OK
 content-length: 8
 
 Hello 1
@@ -39,27 +38,18 @@ content-length: 8
 World 2
 `;
 
-// Do not run this test on windows
-if (Deno.build.os !== "win") {
-  test(async function serverPipelineRace(): Promise<void> {
-    await startServer();
+test(async function serverPipelineRace(): Promise<void> {
+  await startServer();
 
-    nc = run({
-      args: ["nc", "localhost", "4501"],
-      stdin: "piped",
-      stdout: "piped"
-    });
-    const r = new TextProtoReader(new BufReader(nc.stdout));
-    await nc.stdin.write(new TextEncoder().encode(ncIn));
-    const ncOutLines = ncOut.split("\n");
-    // length - 1 to disregard last empty line
-    for (let i = 0; i < ncOutLines.length - 1; i++) {
-      const [s, err] = await r.readLine();
-      assert(!err);
-      assertEquals(s, ncOutLines[i]);
-    }
-    nc.close();
-    nc.stdout.close();
-    killServer();
-  });
-}
+  const conn = await dial("tcp", "127.0.0.1:4501");
+  const r = new TextProtoReader(new BufReader(conn));
+  await conn.write(new TextEncoder().encode(input));
+  const outLines = output.split("\n");
+  // length - 1 to disregard last empty line
+  for (let i = 0; i < outLines.length - 1; i++) {
+    const [s, err] = await r.readLine();
+    assert(!err);
+    assertEquals(s, outLines[i]);
+  }
+  killServer();
+});