--- a/http/test.ts
+++ b/http/test.ts
@@ -1,3 +1,4 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 import "./server_test.ts";
 import "./file_server_test.ts";
+import "./racing_server_test.ts";