+ Using a `Map` to track pipelined requests and resolve only when the previous request is responded.
+ Added test (`racing_server_test.ts`). This test have been confirmed to fail before this patch.


Slight performance impact, yet tail latency is still about the same:

Before:
```
Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   382.03us  216.38us   6.15ms   94.04%
    Req/Sec    13.63k     1.64k   15.48k    86.14%
  274043 requests in 10.10s, 13.07MB read
Requests/sec:  27133.13
Transfer/sec:      1.29MB
```

After:
```
Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   388.00us  176.04us   2.82ms   93.29%
    Req/Sec    13.18k   832.27    14.60k    67.33%
  264768 requests in 10.10s, 12.63MB read
Requests/sec:  26214.26
Transfer/sec:      1.25MB
```