@zekth 2019-04-09
Use `Invoke-WebRequest -Uri "URL"` in Powershell for windows.